requirements.txt
streamlit>=1.32.0
ultralytics>=8.0.0
transformers>=4.38.0
scikit-learn>=1.4.0
pillow>=10.0.0
opencv-python>=4.9.0
numpy>=1.26.0
pandas>=2.0.0
